Move `build_test_artifacts_with_scarb`
