+++ crates/forge/src/scarb_test_collector.rs
@@ -1,6 +1,8 @@
 use crate::collecting::CompiledTestCrateRaw;
-use anyhow::Result;
+use anyhow::{anyhow, Context, Result};
 use camino::Utf8Path;
+use scarb_ui::args::PackagesFilter;
+use std::process::{Command, Stdio};
 
 /// Load test artifacts (deserialized compiled test crates) generated by `scarb snforge-test-collector`
 pub fn load_test_artifacts(
@@ -13,4 +15,20 @@ pub fn load_test_artifacts(
         snforge_test_artifact_path,
     )?)?;
     Ok(test_crates)
+}
+
+pub fn build_test_artifacts_with_scarb(filter: PackagesFilter) -> Result<()> {
+    let build_output = Command::new("scarb")
+        .arg("snforge-test-collector")
+        .env("SCARB_PACKAGES_FILTER", filter.to_env())
+        .stderr(Stdio::inherit())
+        .stdout(Stdio::inherit())
+        .output()
+        .context("Failed to build test artifacts with Scarb")?;
+
+    if build_output.status.success() {
+        Ok(())
+    } else {
+        Err(anyhow!("scarb snforge-test-collector did not succeed"))
+    }
 }
